Media size to Long

+++ src/main/kotlin/core/edconv/data/MediaInfoData.kt
@@ -4,7 +4,7 @@ import kotlinx.serialization.Serializable
 import kotlinx.serialization.json.Json
 
 @Serializable
-data class MediaInfoData(val duration: String, val size: Int) {
+data class MediaInfoData(val duration: String, val size: Long) {
     companion object {
         fun fromJsonString(json: String): MediaInfoData = Json.decodeFromString(json)
     }
